+++ packages/gatsby/cache-dir/static-entry.js
@@ -2,12 +2,13 @@ const React = require(`react`)
 const fs = require(`fs`)
 const { join } = require(`path`)
 const { renderToString, renderToStaticMarkup } = require(`react-dom/server`)
-const { ServerLocation, Router } = require(`@reach/router`)
+const { ServerLocation, Router, isRedirect } = require(`@reach/router`)
 const { get, merge, isObject, flatten, uniqBy } = require(`lodash`)
 
 const apiRunner = require(`./api-runner-ssr`)
 const syncRequires = require(`./sync-requires`)
 const { dataPaths, pages } = require(`./data.json`)
+const { version: gatsbyVersion } = require(`gatsby/package.json`)
 
 // Speed up looking up pages.
 const pagesObjectMap = new Map()
@@ -15,6 +16,10 @@ pages.forEach(p => pagesObjectMap.set(p.path, p))
 
 const stats = JSON.parse(
   fs.readFileSync(`${process.cwd()}/public/webpack.stats.json`, `utf-8`)
+)
+
+const chunkMapping = JSON.parse(
+  fs.readFileSync(`${process.cwd()}/public/chunk-map.json`, `utf-8`)
 )
 
 // const testRequireError = require("./test-require-error")
@@ -46,7 +51,9 @@ const createElement = React.createElement
 
 export default (pagePath, callback) => {
   let bodyHtml = ``
-  let headComponents = []
+  let headComponents = [
+    <meta name="generator" content={`Gatsby ${gatsbyVersion}`} />
+  ]
   let htmlAttributes = {}
   let bodyAttributes = {}
   let preBodyComponents = []
@@ -177,7 +184,12 @@ export default (pagePath, callback) => {
 
   // If no one stepped up, we'll handle it.
   if (!bodyHtml) {
-    bodyHtml = renderToString(bodyComponent)
+    try {
+      bodyHtml = renderToString(bodyComponent)
+    } catch (e) {
+      // ignore @reach/router redirect errors
+      if (!isRedirect(e)) throw e
+    }
   }
 
   // Create paths to scripts
@@ -325,6 +337,21 @@ export default (pagePath, callback) => {
       id={`gatsby-script-loader`}
       dangerouslySetInnerHTML={{
         __html: windowData,
+      }}
+    />
+  )
+
+  // Add chunk mapping metadata
+  const scriptChunkMapping = `/*<![CDATA[*/window.___chunkMapping=${JSON.stringify(
+    chunkMapping
+  )};/*]]>*/`
+
+  postBodyComponents.push(
+    <script
+      key={`chunk-mapping`}
+      id={`gatsby-chunk-mapping`}
+      dangerouslySetInnerHTML={{
+        __html: scriptChunkMapping,
       }}
     />
   )