+++ packages/gatsby/src/utils/api-browser-docs.js
@@ -49,6 +49,14 @@ exports.onRouteUpdateDelayed = true
  * @example
  * exports.onRouteUpdate = ({ location }) => {
  *   console.log('new pathname', location.pathname)
+ *
+ *   // Track pageview with google analytics
+ *   window.ga(
+ *     `set`,
+ *     `page`,
+ *     location.pathname + location.search + location.hash,
+ *   )
+ *   window.ga(`send`, `pageview`)
  * }
  */
 exports.onRouteUpdate = true