+++ packages/gatsby-plugin-guess-js/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="1.0.0"></a>
+
+# [1.0.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-guess-js@1.0.0-rc.1...gatsby-plugin-guess-js@1.0.0) (2018-09-17)
+
+**Note:** Version bump only for package gatsby-plugin-guess-js
+
 <a name="1.0.0-rc.1"></a>
 
 # [1.0.0-rc.1](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-guess-js@1.0.0-rc.0...gatsby-plugin-guess-js@1.0.0-rc.1) (2018-08-29)