+++ packages/gatsby-plugin-typescript/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.0.0"></a>
+
+# [2.0.0](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-typescript/compare/gatsby-plugin-typescript@2.0.0-rc.6...gatsby-plugin-typescript@2.0.0) (2018-09-17)
+
+**Note:** Version bump only for package gatsby-plugin-typescript
+
+<a name="2.0.0-rc.6"></a>
+
+# [2.0.0-rc.6](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-typescript/compare/gatsby-plugin-typescript@2.0.0-rc.5...gatsby-plugin-typescript@2.0.0-rc.6) (2018-09-13)
+
+**Note:** Version bump only for package gatsby-plugin-typescript
+
+<a name="2.0.0-rc.5"></a>
+
+# [2.0.0-rc.5](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-typescript/compare/gatsby-plugin-typescript@2.0.0-rc.4...gatsby-plugin-typescript@2.0.0-rc.5) (2018-09-08)
+
+**Note:** Version bump only for package gatsby-plugin-typescript
+
+<a name="2.0.0-rc.4"></a>
+
+# [2.0.0-rc.4](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-typescript/compare/gatsby-plugin-typescript@2.0.0-rc.3...gatsby-plugin-typescript@2.0.0-rc.4) (2018-09-05)
+
+**Note:** Version bump only for package gatsby-plugin-typescript
+
 <a name="2.0.0-rc.3"></a>
 
 # [2.0.0-rc.3](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-typescript/compare/gatsby-plugin-typescript@2.0.0-rc.2...gatsby-plugin-typescript@2.0.0-rc.3) (2018-08-29)